new syntax for low level matching and term building + delift metas finished
the construction (@HD ARGS) can be used for both matching and
construction.  (ARGS HD@) the same, but here the args are read
right to left, so that one can accumulate [x|ARGS] and then flip
at the last moment.

(#HD ARGS) and (?HD ARGS) are special cases what match only
if the HD is a frozen evar / flexible (respectively).