Automatic GC via Gc.finalise
Note: it does not work as expected. It seems that a few values are
never reclaimed by the GC. Why? Moreover, the finalise function should
remove bindings from a functional data structure that is passed around.
Because we do not have it, we store the bindings to be removed in a list
of things to do and we do that later when possible. However, it seems that
concurrent access to that list is a bad idea as well and some removals
are lost.