Code reindented and mistery solved.
The automatic GC is not working very well.
The reason is that a Var can occur in the codomain of a dead variable.
The major_cycles are the ones that call the finalizers, and they are
very rare. Thus, to remove the variable from the codomain, you need to wait
for a major cycle, followed by a clean, followed by another major cycle
and another clean. To remove a chain of length k (like the ones we have
in the example), you need up to k cycles.

Moreover, a function parameter becomes dead for the GC as soon as it is
dead in the code. However, it seems that a local variable remains alive
until the stack frame of the function/block is reclaimed. Therefore
putting two (Gc.full_major ; clean) one after the other has no effect.

Note: if you cut&paste the commented out code several times, you will
see the size of the bindings go down one by one.