Partial commit: implementation of trail.
NOTE: the commit is partial and the code is bugged because eager
unification + imperative terms is bad. One should either undo and
then redo the bindings all the time (to test eagerly the various
alternatives) or one should pass to the lazy unification. This is what
will try to do next.