Semantics of or changed!
The semantics of Teyjus is not the same as the one of Sicstus that claims
not to be the one of ...

Considering the mess it is, let's pick the one that makes sense in an
ho language. I.e. the ; is just a standard ho predicate implicitly defined
by

; A B :- A.
; A B :- B.

The test has been adjusted accordingly.