Eager + impearive = bad
A warning is issued (misleading because it does not tell you what
implementation is problematic) and an assert false is raised if
the program is non deterministic or if the indexing is not strong enough
to detect that the program is deterministic.

NOTE: the check may slow down a bit non imperative eager unification.