Cut implemented & code refactored.
The new implementation inlines a few functions to avoid
boxing/unboxing.

STRANGE BEHAVIOUR: if you run implementation 18 on rev.elpi
going up from X14 to X18, it always double the time but at
X16 that is almost as fast as X15 (why??). Moreover for X17
and X18 the stack space need to be incremented even if the
code is tail recursive (why??).