Semantics of cut fixed and many more tests for cut added.
NOTE: we used to implement cut using an integer lvl. But
such an implementation is probably not possible because
alternatives with the same lvl could be related to different
choice points. Indeed, both our old implementations and the
original elpi failed on some of the new tests. The new
implementation replaces indexes with pointers to previous
alts (i.e. pointers in the or-list), so that we are sure where
we need to stop (like the WAM does btw).