O(1) unification fragment implemented in 1/3 of the cases.
i.e. in the UVar vs _ case, and not yet in the _ vs (UVar _ | Arg _)

The new test tests/lambda3.ml exercises the new feature on a
type-checker for simply typed lambda calculus. The obtained
algorithm is quadratic in place of linear. Quadratic is to be
expected, but still strange on this particular example where I
supposed the result to be linear. Too much dereferencing, probably.