tests/* ==> source/* and main added
Note (to Enrico): there is no way ATM to automatically test
that the number of results expected is correct and that the
result is really the one we expect up to alpha-equivalence only
(i.e. no unification). The latter could be done with a $foo thing
maybe. Otherwise the Makefile needs to parse the output (looks prone
to continuous changes).