BUG found, fixed, but new bug inserted. All messed up.
When dereferencing a variable when depth > 0 the code was and
still is very bugged. We used to add the depth to the destination,
but that was broken. Now we remove from the source, and it is
broken as well. We need to better understand depth and probably we
need to pass it around as well.