PARTIAL COMMIT: new implementation.
The new implementation is meant to never to_heap the hypotheses of
a clause. The current code should work on the Prolog fragment (but
queens fail, so that must be a bug somewhere).

However, the code is broken for the Lambda-Prolog fragment and it
seem _REALLY_ hard to fix. The point is that now both the r.h.s and
the l.h.s. of unification live in a context that is not the one where
the unification is to be performed. Moreover, the Args occurring in
them live instead in the latter context. This is likely to create a
ton of cases more in unif and to_heap. Do we really want to implement
them?

Performance-wise, on rev this implementation seems to be slower.
Why? (all the call are tail calls)
This implementation enables the other one consisting in not unifying
the heads of the two terms. Before concluding this is slower, we
should try both optimizations together.

PROS: on the other hand, this is the first and only implementation so
far with a nice printing of the result to the user. This was easy
because the initial universally quantified variables are Args and
their instantiation is in an array.