CT: alternative to patternunif (impl 20)
different algorithm for unif/deref/bind that does 1 pass, unfortunately
I'm not sure it it correct (esp delifting).  Performance seem to be ok.
Code is less optimized w.r.t. ==/smart_map but should be more readable.

TODO: not yet "exported" to_heap (should be a call to bind with the right
arguments).