New implementation 20.
NOTE: this is a temporary implementation just to do some
memory benchmarks. It uses the very same algorithm as 18,
but implemented with the data structures of 20. I.e. it passes
around things that are not used or whose value is known in
advance. Preliminary benchmarks show that 20 is already about
10% slower than 18, but still much faster than 19. To be
investigated further.