lambda3 in makam syntax
makam fills my 8G of ram, I had to kill it.

If you run it on a smaller case, then it works, so I guess
it is just slow and memory hog.  As the author told me, it is coded
using "the logic monad" that I'm afraid is like the old epli but with
the extra cost of CPS to code back tracking.  In short, unsure we want
to mention it in the paper