Teyjus turned into a submodule.
When you check out the repository,
remember to run "git submodule init"
and "git submodule update" in every
submodule. And be carefule of the usual
pitfall of undoing commits related to
submodules by not updating the submodule
when the container is updated. To avoid
that, everyone who updates the submodule
_MUST_ explicitly notify the fact to the
developers by mail.