A taste of propagate implemented.
NOTE: the implementation is still very bugged and inefficient, but it
works on our first example (unicity of typing).

BUGS: the constraints are matched only looking at their conclusion and
      using full unification
INEFFICIENCY: we compute all possible pairs containing the new goal;
 we also recompute them as soon as at least one constraints fire.

In the example, we compute pairs/fire many many times.