Pretty-printing code rationalized and bugs fixed
Now the parser exports several precedence levels
used in the runtime.ml code. Moreover, the lambda
case as last argument of an application is now
handled in an ad-hoc case and not via precedences
because it cannot be explained in that way.
Essentially lambda has two distinct precedences
depending on where it occurs.