avoid messing up the term during pretty-printing
The invariants for heap terms do not allow to pretty-print a term
dereferencing its Args. The reason is that the args do not live at the
depth of the term, but at an higher depth. Dereferencing them at the
depth we have in the function triggers wrong restrictions and imperatively
messes up the term forever.