WARNING: HACKY COMMIT TO BE REVERTED EVENTUALLY
We literally delay hundreds of goals. Until we find a way to declare
proper propagation rules, this patch removes duplicates, defined as
goals that are printed in the same way.... (because the terms are cyclic,
comparing two term diverges)

With this patch we still have tens of goals that should be propagated.