thenll! removed in favour of ! that allows to implement orelse! as well
NOTE: left still allows backtracking because banging (mem Gamma X) is
useless because then the script looks for X again. A good reason to
let mem retrieve the path to the hypothesis to be used in the script.
Or, maybe better, to have mem move the hypothesis to the stoup.