functions (matching) also used to propagate
1.Syntax to declare a block of functions, one local.

  function { local aux-c.
    aux-c (?X L) :- X = (x\3), L = [4].
    c X Y :- aux-c X, Y = b.
  }

c and aux-c must have linear vars in the head (ceap check at runtime
also on dynamic clauses, that could not be easily checked statically),
and such heads are *matched*, not unified.  As a bonus (?HD ARGS) matches meta-level
non-instantiated unification variable (and its arguments) providing a form
of introspection ($is_flex, ...).   Also ? and (? HD) or ?HD are valid
patterns to match an uvar: the first does not bind anything, the latter 2
do only bind the head.
Final bonus, the "as X" caml syntax was added.  The point is that
re-building a meta after pattern matching is hard, so

  term (? as X) :- $delay (term X)

does what you expect.  as works in any position, so one can write

  rev ([X|XS] as In) ... :- $print "reversing " In, ...

2.Desugaring, to help linearize, -> and := as follows

  c X -> t :- b.    ====    c X Frehs :- Fresh = t, b.

  c [X|XS] -> (f Y) :- Y := c XS.   =====
    c [X|XS] Fresh :- Fresh = f Y, c XS Y.

note that := is there only to help the eye, since the head looks like
a clause with one argumet less, so also the rec-call should.

3.Example: Never generative copy function

function {
  copy (app X XS) -> (app Y YS) :- Y := copy X, YS := copy XS.
  copy (lam F) -> (lam G) :- pi x\ copy x -> x => G x := copy (F x).
  copy c -> c.
}

main :-
  (copy X Y, $print "gen mode is wrong", !, false);
  (copy (lam x\app (W x) c) R, $print "rec gen mode is wrong", !, false);
  ($print "ok", copy (lam x\app x c) R, $print R).

4. Constraints are grouped in cliques and each clique has

constraint c1 c2 c3 {
  rule [ M ] [ R ] -> G :- cond.
  rule ...
}

Here M and R are matched.  {c1,c2,c3} is a clique, and the rules
for the clique only apply to constraints that have c1 or c2 or c3 as head.
Also,  $constrain to delay filtering the context according to
the clique the constraint lives in.

5. pruning duplicate constraints in CHR now works.  Eg: term.elpi

TODO: take advantage of linearity + data-driven to skip occur-check
      if the clause validated the rules of Apt's paper/
      CFR: branch claudio

TODO: take advantage of the closed tag in propagate/freeze by
      exposing it as a hard constant.  Brainstorm:

  rule [] [ append (closed L) (closed R) O ] -> (append L R (closed O)).

Conflicts:
	elpi/elpi_custom.ml
	elpi/elpi_runtime.ml
	elpi/elpi_runtime.mli
	elpi/elpi_trace.ml