test for custom contraint (coq universes checker)
- minimal import of Coq code for universes
- declaration of universe constraints and universe data type
- check that backtracking works

The test is not very deep but checks that = can be extended
using the semantics of the custom objects: two constraints
are ground terms that compare equal only if such constraint
is in the store.  This means that one cannot make them be equal
mistake (passing arguments to prolog functions) but that one
can still use = to /check/ if they are already equal.