in (?? K L) K is not first class: only $assign can assign K
$assign does no check, but could in principle verify the unifyer is "well scoped".

I see a HUGE problem here.  Imagine "0 * x" assigned to "K y".  This is valid
in CIC, but not in LP.  It is the converse problem of lambdas
at the metal level or at the object language level.  In Matita
"X 3" and "X true" have no solution for X, while in LP "x\Y" is.