Refinement fix (partial): split on meta during on_match_type
Here and in at least a couple more places, split can be performed
over the (empty, or too short) stack of a meta.

This commit fixes one case, but:
1) it counts the arguments of an inductive not up-to-beta
   (to fix, Ferruccio maybe?)
2) the code should be generalized and re-used in all places where
   split is used

Even with this patch, the example I was looking for does not pass
(it diverges). However, I suspect that divergence is caused by
something else.