during projection, conversion should precede instantiation
since instantiation is wrong if conversion fails,
and we can produce a lot of backtraking due to contraints on the meta woken-up by the instantiation.

We use the custom predicate $occurs to assert if projection is possible,
so to rule out useless conversions before the instantiation.

now the 4th refinement problem is solved in the way we expect